LayoutTests/http/tests/security/module-no-mime-type.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test module rejects scripts with no mime type.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script>
debug('Module is not executed yet.');
function done()
{
    debug('Module rejects a script with no mime type.');
}
window.scriptsSuccessfullyLoaded = 0;
</script>
<script src="../../resources/js-test-post.js"></script>
<script type="module" src="./contentTypeOptions/resources/script-with-header.pl" onerror="done()"></script>
<script type="module">
shouldBe(`window.scriptsSuccessfullyLoaded`, `0`);
finishJSTest();
</script>
</body>
</html>
